@@ -10,21 +10,26 @@
     "@tuleap/fp-feedback": "workspace:*",
     "@tuleap/list-picker": "workspace:*",
     "@tuleap/tlp-fetch": "workspace:*",
-    "@tuleap/vue2-gettext-init": "workspace:*",
-    "vue": "^2.7.8",
-    "vuex": "^3.6.2"
+    "@tuleap/vue-strict-inject": "workspace:*",
+    "@tuleap/vue3-gettext-init": "workspace:*",
+    "vue": "^3.3.4",
+    "vue3-gettext": "^2.4.0",
+    "vuex": "^4.1.0",
+    "vuex-composition-helpers": "^2.0.1"
   },
   "devDependencies": {
     "@tuleap/build-system-configurator": "workspace:*",
     "@tuleap/core-constants": "workspace:*",
     "@tuleap/po-gettext-plugin": "workspace:*",
     "@tuleap/vuex-store-wrapper-jest": "workspace:^",
-    "@vue/test-utils": "^1.3",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "@vue/test-utils": "^2.2.10",
     "jquery": "3.6.0"
   },
   "scripts": {
-    "build": "webpack --config webpack.prod.js",
-    "watch": "webpack --config webpack.dev.js --watch",
-    "test": "jest"
+    "typecheck": "vue-tsc --noEmit",
+    "build": "vite build",
+    "watch": "NODE_ENV=development vite build --watch --mode development --minify false",
+    "test": "vitest"
   }
 }
