@@ -17,15 +17,29 @@
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import type { Vue } from "vue/types/vue";
-import { createLocalVue } from "@vue/test-utils";
-import { initVueGettextFromPoGettextPlugin } from "@tuleap/vue2-gettext-init";
+import { describe, it, expect } from "vitest";
+import type { Project } from "../store/types";
+import { ProjectsSorter } from "./ProjectsSorter";
 
-export async function createMoveModalLocalVue(): Promise<typeof Vue> {
-    const local_vue = createLocalVue();
-    await initVueGettextFromPoGettextPlugin(local_vue, () => {
-        throw new Error("Fallback to default");
+describe("ProjectsSorter", () => {
+    it("sorted_projects should return the projects alphabetically sorted", () => {
+        const projects: Project[] = [
+            {
+                id: 105,
+                label: "Scrum",
+            },
+            {
+                id: 106,
+                label: "Git",
+            },
+            {
+                id: 107,
+                label: "Kanban",
+            },
+        ];
+
+        expect(
+            ProjectsSorter.sortProjectsAlphabetically(projects).map(({ id }) => id)
+        ).toStrictEqual([106, 107, 105]);
     });
-
-    return local_vue;
-}
+});
