@@ -1,5 +1,5 @@
 /*
- * Copyright (c) Enalean, 2020-Present. All Rights Reserved.
+ * Copyright (c) Enalean, 2023 - present. All Rights Reserved.
  *
  * This file is a part of Tuleap.
  *
@@ -17,7 +17,10 @@
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import common from "./webpack.common.js";
-import { webpack_configurator } from "@tuleap/build-system-configurator";
+import type { Project } from "../store/types";
 
-export default webpack_configurator.extendProdConfiguration(common);
+export const ProjectsSorter = {
+    sortProjectsAlphabetically(projects: Project[]): Project[] {
+        return projects.sort((a: Project, b: Project) => a.label.localeCompare(b.label));
+    },
+};
