@@ -17,22 +17,17 @@
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import Vue from "vue";
-import Vuex from "vuex";
+import { createStore } from "vuex";
 import type { Store } from "vuex";
 import { default_state } from "./state";
 import * as actions from "./actions";
 import * as mutations from "./mutations";
-import * as getters from "./getters";
 import type { RootState } from "./types";
 
-Vue.use(Vuex);
-
-export function createStore(): Store<RootState> {
-    return new Vuex.Store({
+export function createInitializedStore(): Store<RootState> {
+    return createStore({
         state: { ...default_state },
         actions,
         mutations,
-        getters,
     });
 }
