@@ -18,6 +18,7 @@
  */
 
 declare module "*.vue" {
-    import Vue from "vue";
-    export default Vue;
+    import type { DefineComponent } from "vue";
+    const component: DefineComponent;
+    export default component;
 }
